pyproject.toml
[tool.poetry]
name = "umn-ta2-database-processing"
version = "0.2.0"
description = ""
authors = ["Jiyoon Pyo <jiyoonp0228@gmail.com>"]
license = "MIT"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
beautifulsoup4 = "^4.9.1"
geopandas = "^0.13.2"
numpy = "^1.22.1"
pyarrow = "^14.0.1"
polars = "^0.19.19"
regex = "^2023.10.3"
networkx = "^3.1"
scipy = "^1.10.1"
huggingface-hub = "^0.20.3"
pandas = "^2.0.3"
requests = "^2.28.1"
scikit-learn = "^1.3.1"
sentence-transformers = "^2.2.2"
sentencepiece = "^0.1.99"
spacy = "^3.7.4"
torch = "^2.5.1"
torchvision = "^0.20.1"
torchaudio = "^2.5.1"
en-core-web-sm = {url = "https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889"}
strsim-2 = "^1.1.2"
tabulate = "^0.9.0"
torchmetrics = "^1.4.0"
tqdm = "^4.66.2"
transformers = "^4.33.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
